tools/lit-builder/pyproject.toml
[project]
name = "lit-builder"
version = "0.1.0"
description = "ICLR/NeurIPS literature builder with keyword + LLM relevance scoring"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "anthropic>=0.39",
    "ollama>=0.4",
    "rich>=13.7",
    "platformdirs>=4.2",
]

[project.optional-dependencies]
docs = ["mkdocs>=1.6", "mkdocs-material>=9.5"]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
lit = "lit_builder.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lit_builder"]

[tool.ruff]
line-length = 100
target-version = "py311"
